requirements.txt
tkinter
matplotlib
python-dotenv
google-generativeai
